Refactor helpers into dedicated files
