+++ includes/functions.php
@@ -120,113 +120,6 @@ function hic_reliable_polling_enabled() { return hic_get_option('reliable_pollin
 
 // Admin and General Settings
 function hic_get_admin_email() { return hic_get_option('admin_email', get_option('admin_email')); }
-function hic_get_log_file() {
-    return hic_get_option('log_file', WP_CONTENT_DIR . '/uploads/hic-logs/hic-log.txt');
-}
-
-function hic_validate_log_path($path) {
-    $base_dir = WP_CONTENT_DIR . '/uploads/hic-logs/';
-    $default = $base_dir . 'hic-log.txt';
-
-    $path = sanitize_text_field($path);
-    if (empty($path)) {
-        return $default;
-    }
-
-    $normalized_path = str_replace('\\', '/', $path);
-    $normalized_base = rtrim(str_replace('\\', '/', $base_dir), '/') . '/';
-
-    if (strpos($normalized_path, '..') !== false || strpos($normalized_path, $normalized_base) !== 0) {
-        return $default;
-    }
-
-    return $normalized_path;
-}
-
-/**
- * Mask common sensitive data like emails, phone numbers and tokens.
- */
-function hic_mask_sensitive_data($message) {
-    if (!is_string($message)) {
-        return $message;
-    }
-
-    // Mask email addresses
-    $message = preg_replace('/[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}/u', '[masked-email]', $message);
-
-    // Mask phone numbers (sequences of digits, spaces or dashes)
-    $message = preg_replace('/\b(?:\+?\d[\d\s\-]{7,}\d)\b/u', '[masked-phone]', $message);
-
-    // Mask tokens and authorization headers
-    $message = preg_replace('/(token|api[_-]?key|secret|password)\s*[=:]\s*[A-Za-z0-9._-]+/iu', '$1=[masked]', $message);
-    $message = preg_replace('/Authorization:\s*Bearer\s+[A-Za-z0-9._-]+/iu', 'Authorization: Bearer [masked]', $message);
-
-    return $message;
-}
-
-/**
- * Default filter for hic_log_message that masks sensitive data.
- *
- * @param string $message Log message
- * @param string $level   Log level
- * @return string
- */
-function hic_default_log_message_filter($message, $level) {
-    if (is_array($message)) {
-        foreach ($message as $key => $value) {
-            if (is_string($key) && preg_match('/^(?:email)$/i', $key)) {
-                $message[$key] = '[masked-email]';
-                continue;
-            }
-            if (is_string($key) && preg_match('/(token|api[_-]?key|secret|password)/i', $key)) {
-                $message[$key] = '[masked]';
-                continue;
-            }
-            $message[$key] = hic_default_log_message_filter($value, $level);
-        }
-        return $message;
-    }
-
-    if (is_object($message)) {
-        foreach ($message as $key => $value) {
-            if (is_string($key) && preg_match('/^(?:email)$/i', $key)) {
-                $message->$key = '[masked-email]';
-                continue;
-            }
-            if (is_string($key) && preg_match('/(token|api[_-]?key|secret|password)/i', $key)) {
-                $message->$key = '[masked]';
-                continue;
-            }
-            $message->$key = hic_default_log_message_filter($value, $level);
-        }
-        return $message;
-    }
-
-    if (is_bool($message) || $message === null) {
-        return $message;
-    }
-
-    if (is_string($message)) {
-        $masked = hic_mask_sensitive_data($message);
-        if ($masked !== $message) {
-            return $masked;
-        }
-        if (is_numeric($message)) {
-            return '[masked-number]';
-        }
-        return $masked;
-    }
-
-    if (is_numeric($message)) {
-        return '[masked-number]';
-    }
-
-    return $message;
-}
-
-if (function_exists('add_filter')) {
-    add_filter('hic_log_message', __NAMESPACE__ . '\\hic_default_log_message_filter', 10, 2);
-}
 
 // GTM Settings
 function hic_is_gtm_enabled() { return hic_get_option('gtm_enabled', '0') === '1'; }
@@ -315,280 +208,7 @@ function hic_release_polling_lock() {
 /**
  * Check if retry event should be scheduled based on conditions
  */
-function hic_should_schedule_retry_event() {
-    if (!hic_realtime_brevo_sync_enabled()) {
-        return false;
-    }
-    
-    if (!hic_get_brevo_api_key()) {
-        return false;
-    }
-    
-    $schedules = wp_get_schedules();
-    return isset($schedules['hic_retry_interval']);
-}
 
-/* ============ Privacy Exporter/Eraser ============ */
-
-function hic_register_exporter($exporters) {
-    $exporters['hic-tracking-data'] = [
-        'exporter_friendly_name' => __('Dati di tracciamento HIC', 'hotel-in-cloud'),
-        'callback' => __NAMESPACE__ . '\\hic_export_tracking_data',
-    ];
-    return $exporters;
-}
-
-function hic_register_eraser($erasers) {
-    $erasers['hic-tracking-data'] = [
-        'eraser_friendly_name' => __('Dati di tracciamento HIC', 'hotel-in-cloud'),
-        'callback' => __NAMESPACE__ . '\\hic_erase_tracking_data',
-    ];
-    return $erasers;
-}
-
-/**
- * Export tracking data associated with an email address.
- */
-function hic_export_tracking_data($email_address, $page = 1) {
-    global $wpdb;
-
-    if (!$wpdb) {
-        return ['data' => [], 'done' => true];
-    }
-
-    $table = $wpdb->prefix . 'hic_gclids';
-    $table_exists = $wpdb->get_var($wpdb->prepare("SHOW TABLES LIKE %s", $table)) === $table;
-    if (!$table_exists) {
-        return ['data' => [], 'done' => true];
-    }
-
-    $email = sanitize_email($email_address);
-    if (empty($email)) {
-        return ['data' => [], 'done' => true];
-    }
-
-    $column_exists = $wpdb->get_var($wpdb->prepare("SHOW COLUMNS FROM $table LIKE %s", 'email'));
-    if (!$column_exists) {
-        return ['data' => [], 'done' => true];
-    }
-
-    $number = 50;
-    $offset = ($page - 1) * $number;
-    $rows = $wpdb->get_results(
-        $wpdb->prepare(
-            "SELECT id, gclid, fbclid, sid, utm_source, utm_medium, utm_campaign, utm_content, utm_term, created_at FROM $table WHERE email = %s ORDER BY id ASC LIMIT %d OFFSET %d",
-            $email,
-            $number,
-            $offset
-        )
-    );
-
-    $data = [];
-    foreach ($rows as $row) {
-        $item = [];
-        if ($row->gclid !== null) {
-            $item[] = ['name' => 'gclid', 'value' => $row->gclid];
-        }
-        if ($row->fbclid !== null) {
-            $item[] = ['name' => 'fbclid', 'value' => $row->fbclid];
-        }
-        if ($row->sid !== null) {
-            $item[] = ['name' => 'sid', 'value' => $row->sid];
-        }
-        if ($row->utm_source !== null) {
-            $item[] = ['name' => 'utm_source', 'value' => $row->utm_source];
-        }
-        if ($row->utm_medium !== null) {
-            $item[] = ['name' => 'utm_medium', 'value' => $row->utm_medium];
-        }
-        if ($row->utm_campaign !== null) {
-            $item[] = ['name' => 'utm_campaign', 'value' => $row->utm_campaign];
-        }
-        if ($row->utm_content !== null) {
-            $item[] = ['name' => 'utm_content', 'value' => $row->utm_content];
-        }
-        if ($row->utm_term !== null) {
-            $item[] = ['name' => 'utm_term', 'value' => $row->utm_term];
-        }
-        if ($row->created_at !== null) {
-            $item[] = ['name' => 'created_at', 'value' => $row->created_at];
-        }
-
-        $data[] = [
-            'group_id'    => 'hic_tracking_data',
-            'group_label' => __('Dati di tracciamento HIC', 'hotel-in-cloud'),
-            'item_id'     => 'tracking-' . $row->id,
-            'data'        => $item,
-        ];
-    }
-
-    $done = count($rows) < $number;
-
-    return ['data' => $data, 'done' => $done];
-}
-
-/**
- * Erase tracking data associated with an email address.
- */
-function hic_erase_tracking_data($email_address, $page = 1) {
-    global $wpdb;
-
-    if (!$wpdb) {
-        return ['items_removed' => false, 'items_retained' => false, 'messages' => [], 'done' => true];
-    }
-
-    $table = $wpdb->prefix . 'hic_gclids';
-    $table_exists = $wpdb->get_var($wpdb->prepare("SHOW TABLES LIKE %s", $table)) === $table;
-    if (!$table_exists) {
-        return ['items_removed' => false, 'items_retained' => false, 'messages' => [], 'done' => true];
-    }
-
-    $email = sanitize_email($email_address);
-    if (empty($email)) {
-        return ['items_removed' => false, 'items_retained' => false, 'messages' => [], 'done' => true];
-    }
-
-    $column_exists = $wpdb->get_var($wpdb->prepare("SHOW COLUMNS FROM $table LIKE %s", 'email'));
-    if (!$column_exists) {
-        return ['items_removed' => false, 'items_retained' => false, 'messages' => [], 'done' => true];
-    }
-
-    $number = 50;
-    $offset = ($page - 1) * $number;
-    $rows = $wpdb->get_results(
-        $wpdb->prepare(
-            "SELECT id FROM $table WHERE email = %s ORDER BY id ASC LIMIT %d OFFSET %d",
-            $email,
-            $number,
-            $offset
-        )
-    );
-
-    $items_removed = false;
-    foreach ($rows as $row) {
-        $deleted = $wpdb->delete($table, ['id' => $row->id], ['%d']);
-        if ($deleted) {
-            $items_removed = true;
-        }
-    }
-
-    $done = count($rows) < $number;
-
-    return [
-        'items_removed'  => $items_removed,
-        'items_retained' => false,
-        'messages'       => [],
-        'done'           => $done,
-    ];
-}
-
-// Note: These hooks are now registered in hic_init_helper_hooks() function
-// add_filter('wp_privacy_personal_data_exporters', __NAMESPACE__ . '\\hic_register_exporter');
-// add_filter('wp_privacy_personal_data_erasers', __NAMESPACE__ . '\\hic_register_eraser');
-
-/* ============ New Helper Functions ============ */
-
-/**
- * Retrieve tracking IDs (gclid and fbclid) for a given SID from database.
- *
- * @param string $sid Session identifier
- * @return array{gclid:?string, fbclid:?string, msclkid:?string, ttclid:?string}
-*/
-function hic_get_tracking_ids_by_sid($sid) {
-    static $cache = [];
-    $sid = sanitize_text_field($sid);
-    if (empty($sid)) {
-        return ['gclid' => null, 'fbclid' => null, 'msclkid' => null, 'ttclid' => null];
-    }
-
-    if (array_key_exists($sid, $cache)) {
-        return $cache[$sid];
-    }
-
-    global $wpdb;
-    if (!$wpdb) {
-        hic_log('hic_get_tracking_ids_by_sid: wpdb is not available');
-        return $cache[$sid] = ['gclid' => null, 'fbclid' => null, 'msclkid' => null, 'ttclid' => null];
-    }
-
-    $table = $wpdb->prefix . 'hic_gclids';
-
-    // Ensure table exists before querying
-    $table_exists = $wpdb->get_var($wpdb->prepare("SHOW TABLES LIKE %s", $table)) === $table;
-    if (!$table_exists) {
-        hic_log('hic_get_tracking_ids_by_sid: Table does not exist: ' . $table);
-        return $cache[$sid] = ['gclid' => null, 'fbclid' => null, 'msclkid' => null, 'ttclid' => null];
-    }
-    $row = $wpdb->get_row($wpdb->prepare("SELECT gclid, fbclid, msclkid, ttclid FROM $table WHERE sid=%s ORDER BY id DESC LIMIT 1", $sid));
-
-    if ($wpdb->last_error) {
-        hic_log('hic_get_tracking_ids_by_sid: Database error retrieving tracking IDs: ' . $wpdb->last_error);
-        return $cache[$sid] = ['gclid' => null, 'fbclid' => null, 'msclkid' => null, 'ttclid' => null];
-    }
-
-    if ($row) {
-        return $cache[$sid] = [
-            'gclid' => $row->gclid,
-            'fbclid' => $row->fbclid,
-            'msclkid' => $row->msclkid,
-            'ttclid' => $row->ttclid,
-        ];
-    }
-
-    return $cache[$sid] = ['gclid' => null, 'fbclid' => null, 'msclkid' => null, 'ttclid' => null];
-}
-
-/**
- * Retrieve UTM parameters for a given SID from database.
- *
- * @param string $sid Session identifier
- * @return array{utm_source:?string, utm_medium:?string, utm_campaign:?string, utm_content:?string, utm_term:?string}
-*/
-function hic_get_utm_params_by_sid($sid) {
-    static $cache = [];
-    $sid = sanitize_text_field($sid);
-    if (empty($sid)) {
-        return ['utm_source' => null, 'utm_medium' => null, 'utm_campaign' => null, 'utm_content' => null, 'utm_term' => null];
-    }
-
-    if (array_key_exists($sid, $cache)) {
-        return $cache[$sid];
-    }
-
-    global $wpdb;
-    if (!$wpdb) {
-        hic_log('hic_get_utm_params_by_sid: wpdb is not available');
-        return $cache[$sid] = ['utm_source' => null, 'utm_medium' => null, 'utm_campaign' => null, 'utm_content' => null, 'utm_term' => null];
-    }
-
-    $table = $wpdb->prefix . 'hic_gclids';
-
-    // Ensure table exists before querying
-    $table_exists = $wpdb->get_var($wpdb->prepare("SHOW TABLES LIKE %s", $table)) === $table;
-    if (!$table_exists) {
-        hic_log('hic_get_utm_params_by_sid: Table does not exist: ' . $table);
-        return $cache[$sid] = ['utm_source' => null, 'utm_medium' => null, 'utm_campaign' => null, 'utm_content' => null, 'utm_term' => null];
-    }
-
-    $row = $wpdb->get_row($wpdb->prepare("SELECT utm_source, utm_medium, utm_campaign, utm_content, utm_term FROM $table WHERE sid=%s ORDER BY id DESC LIMIT 1", $sid));
-
-    if ($wpdb->last_error) {
-        hic_log('hic_get_utm_params_by_sid: Database error retrieving UTM params: ' . $wpdb->last_error);
-        return $cache[$sid] = ['utm_source' => null, 'utm_medium' => null, 'utm_campaign' => null, 'utm_content' => null, 'utm_term' => null];
-    }
-
-    if ($row) {
-        return $cache[$sid] = [
-            'utm_source'   => $row->utm_source,
-            'utm_medium'   => $row->utm_medium,
-            'utm_campaign' => $row->utm_campaign,
-            'utm_content'  => $row->utm_content,
-            'utm_term'     => $row->utm_term,
-        ];
-    }
-
-    return $cache[$sid] = ['utm_source' => null, 'utm_medium' => null, 'utm_campaign' => null, 'utm_content' => null, 'utm_term' => null];
-}
 
 function hic_normalize_price($value) {
     if (empty($value) || (!is_numeric($value) && !is_string($value))) return 0.0;
@@ -786,40 +406,10 @@ function hic_store_failed_request($url, $args, $error) {
             'payload'    => wp_json_encode($args),
             'attempts'   => 1,
             'last_error' => $error,
-            'last_try'   => current_time('mysql'),
+        'last_try'   => current_time('mysql'),
         ],
         ['%s', '%s', '%d', '%s', '%s']
     );
-}
-
-function hic_log($msg, $level = HIC_LOG_LEVEL_INFO, $context = []) {
-    $log_manager = function_exists('\\hic_get_log_manager') ? \hic_get_log_manager() : null;
-
-    if ($log_manager) {
-        return $log_manager->log($msg, $level, $context);
-    }
-    return false;
-}
-
-/**
- * Normalize bucket attribution according to priority: gclid > fbclid > organic
- * 
- * @param string|null $gclid Google Click ID from Google Ads
- * @param string|null $fbclid Facebook Click ID from Meta Ads
- * @return string One of: 'gads', 'fbads', 'organic'
- */
-function fp_normalize_bucket($gclid, $fbclid){
-  if (!empty($gclid) && trim($gclid) !== '')  return 'gads';
-  if (!empty($fbclid) && trim($fbclid) !== '') return 'fbads';
-  return 'organic';
-}
-
-/**
- * Legacy function name for backward compatibility
- * @deprecated Use fp_normalize_bucket() instead
- */
-function hic_get_bucket($gclid, $fbclid){
-  return fp_normalize_bucket($gclid, $fbclid);
 }
 
 /* ============ Email admin (include bucket) ============ */
@@ -1299,157 +889,6 @@ function hic_get_processing_statistics() {
     return $statistics;
 }
 
-/* ================= SAFE WORDPRESS CRON HELPERS ================= */
-
-/**
- * Safely check if an event is scheduled
- */
-function hic_safe_wp_next_scheduled($hook) {
-    if (!function_exists('wp_next_scheduled')) {
-        return false;
-    }
-    return wp_next_scheduled($hook);
-}
-
-/**
- * Safely schedule an event
- */
-function hic_safe_wp_schedule_event($timestamp, $recurrence, $hook, $args = array()) {
-    if (!function_exists('wp_schedule_event')) {
-        return false;
-    }
-    $result = wp_schedule_event($timestamp, $recurrence, $hook, $args, true);
-    if (is_wp_error($result)) {
-        hic_log('Scheduling error for ' . $hook . ': ' . $result->get_error_message(), HIC_LOG_LEVEL_ERROR);
-        return false;
-    }
-    return $result;
-}
-
-/**
- * Safely clear scheduled hooks
- */
-function hic_safe_wp_clear_scheduled_hook($hook, $args = array()) {
-    if (!function_exists('wp_clear_scheduled_hook')) {
-        return false;
-    }
-    return wp_clear_scheduled_hook($hook, $args);
-}
-
-function hic_add_failed_request_schedule($schedules) {
-    $schedules['hic_every_fifteen_minutes'] = array(
-        'interval' => 15 * 60,
-        'display'  => 'Every 15 Minutes (HIC Failed Requests)'
-    );
-    return $schedules;
-}
-// Note: This hook is now registered in hic_init_helper_hooks() function
-// add_filter('cron_schedules', __NAMESPACE__ . '\\hic_add_failed_request_schedule');
-
-function hic_schedule_failed_request_retry() {
-    if (!hic_safe_wp_next_scheduled('hic_retry_failed_requests')) {
-        hic_safe_wp_schedule_event(time(), 'hic_every_fifteen_minutes', 'hic_retry_failed_requests');
-    }
-}
-// Note: This hook is now registered in hic_init_helper_hooks() function
-// add_action('init', __NAMESPACE__ . '\\hic_schedule_failed_request_retry');
-
-function hic_retry_failed_requests() {
-    global $wpdb;
-    if (!$wpdb) {
-        return;
-    }
-
-    $table = esc_sql($wpdb->prefix . 'hic_failed_requests');
-    $rows  = $wpdb->get_results("SELECT * FROM $table");
-
-    if (empty($rows)) {
-        return;
-    }
-
-    foreach ($rows as $row) {
-        if ($row->attempts >= 5) {
-            $wpdb->delete($table, array('id' => $row->id));
-            continue;
-        }
-
-        $delay     = 15 * 60 * pow(2, max(0, $row->attempts - 1));
-        $next_time = strtotime($row->last_try) + $delay;
-        if (time() < $next_time) {
-            continue;
-        }
-
-        $args = json_decode($row->payload, true);
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            $error_message = 'JSON decode error: ' . json_last_error_msg();
-            hic_log('Retry failed for ' . $row->endpoint . ': ' . $error_message, HIC_LOG_LEVEL_ERROR);
-            $wpdb->delete($table, array('id' => $row->id));
-            continue;
-        }
-        $response = hic_http_request($row->endpoint, is_array($args) ? $args : array());
-
-        if (is_wp_error($response) || wp_remote_retrieve_response_code($response) >= 400) {
-            $error_message = is_wp_error($response) ? $response->get_error_message() : 'HTTP ' . wp_remote_retrieve_response_code($response);
-            $wpdb->update(
-                $table,
-                array(
-                    'attempts'   => $row->attempts + 1,
-                    'last_error' => $error_message,
-                    'last_try'   => current_time('mysql'),
-                ),
-                array('id' => $row->id),
-                array('%d', '%s', '%s'),
-                array('%d')
-            );
-            hic_log('Retry failed for ' . $row->endpoint . ': ' . $error_message, HIC_LOG_LEVEL_ERROR);
-            if ($row->attempts + 1 >= 5) {
-                $wpdb->delete($table, array('id' => $row->id));
-            }
-        } else {
-            hic_log('Retry succeeded for ' . $row->endpoint);
-            $wpdb->delete($table, array('id' => $row->id));
-        }
-    }
-}
-
-function hic_cleanup_failed_requests($days = 30) {
-    if ($days <= 0) {
-        return 0;
-    }
-
-    global $wpdb;
-    if (!$wpdb) {
-        hic_log('hic_cleanup_failed_requests: wpdb is not available');
-        return false;
-    }
-
-    $table  = $wpdb->prefix . 'hic_failed_requests';
-    $cutoff = gmdate('Y-m-d H:i:s', strtotime("-{$days} days"));
-
-    $deleted = $wpdb->query(
-        $wpdb->prepare(
-            "DELETE FROM $table WHERE last_try < %s",
-            $cutoff
-        )
-    );
-
-    if ($deleted === false) {
-        hic_log('hic_cleanup_failed_requests: Database error: ' . $wpdb->last_error);
-        return false;
-    }
-
-    hic_log("hic_cleanup_failed_requests: Removed $deleted records older than $days days");
-    return $deleted;
-}
-
-function hic_schedule_failed_request_cleanup() {
-    if (!hic_safe_wp_next_scheduled('hic_cleanup_failed_requests')) {
-        hic_safe_wp_schedule_event(time(), 'daily', 'hic_cleanup_failed_requests');
-    }
-}
-// Note: This hook is now registered in hic_init_helper_hooks() function
-// add_action('hic_retry_failed_requests', __NAMESPACE__ . '\\hic_retry_failed_requests');
-// add_action('hic_cleanup_failed_requests', __NAMESPACE__ . '\\hic_cleanup_failed_requests');
 
 }
